separated validation.js into constants, functions, and schema. Next will divide schema into its own folder
