+++ client/src/constants/validation/constants.js
@@ -1,0 +1,23 @@
+export const healthRegions = [
+  'Interior',
+  'Fraser',
+  'Vancouver Coastal',
+  'Vancouver Island',
+  'Northern',
+];
+
+export const foundOutReasons = [
+  'Friend(s)',
+  'WorkBC',
+  'Governement announcement',
+  'Colleague(s)',
+  'Job posting through Health Authority',
+  'Job posting with employer',
+  'Web search',
+  'Social media',
+  'Other',
+];
+
+export const siteTypes = ['Long-term care', 'Assisted living', 'Both', 'Other'];
+
+export const userRoles = ['health_authority', 'employer', 'ministry_of_health'];
