+++ client/src/constants/validation/schema.js
@@ -1,122 +1,23 @@
 import * as yup from 'yup';
 import mapValues from 'lodash/mapValues';
+import { healthRegions, foundOutReasons, siteTypes, userRoles } from './constants';
 import {
-  archiveReasonOptions,
-  archiveStatusOptions,
-  minDateString,
-  maxDateString,
-} from './archiveParticipantsConstants';
-import { indigenousIdentities } from '../components/modal-forms/IndigenousDeclarationForm';
+  validateDateString,
+  validatePastDateString,
+  validateDateIsInThePast,
+  validateDateIsReasonable,
+  validateUniqueArray,
+  validateBlankOrPositiveInteger,
+  errorMessage,
+} from './functions';
+import { archiveReasonOptions, archiveStatusOptions } from '../archiveParticipantsConstants';
+import { indigenousIdentities } from '../../components/modal-forms/IndigenousDeclarationForm';
+import { postHireStatuses, postHireStatusesValues } from '../../constants/postHireConstants';
+import { rosEmploymentTypeValues, rosPositionTypeValues } from '../return-of-service';
 
-import { postHireStatuses, postHireStatusesValues } from '../constants/postHireConstants';
-import { rosEmploymentTypeValues, rosPositionTypeValues } from './return-of-service';
-
-const healthRegions = ['Interior', 'Fraser', 'Vancouver Coastal', 'Vancouver Island', 'Northern'];
-
-const foundOutReasons = [
-  'Friend(s)',
-  'WorkBC',
-  'Governement announcement',
-  'Colleague(s)',
-  'Job posting through Health Authority',
-  'Job posting with employer',
-  'Web search',
-  'Social media',
-  'Other',
-];
-
-const siteTypes = ['Long-term care', 'Assisted living', 'Both', 'Other'];
-
-const userRoles = ['health_authority', 'employer', 'ministry_of_health'];
-
-const validateDateString = (s) => {
-  if (/^\d{4}\/\d{2}\/\d{2}$/.test(s) === false) return false;
-  const date = Date.parse(s);
-  return typeof date === 'number' && !Number.isNaN(date);
-};
-
-const validatePastDateString = (s) => {
-  if (!validateDateString(s)) return false;
-  return Date.parse(s) <= new Date();
-};
-const validateDateIsInThePast = (s) => {
-  return Date.parse(s) <= new Date();
-};
-const validateDateIsReasonable = (d) => {
-  try {
-    return Date.parse(d) >= Date.parse(minDateString) && Date.parse(d) <= Date.parse(maxDateString);
-  } catch (e) {
-    return false;
-  }
-};
-const validateUniqueArray = (a) => Array.isArray(a) && new Set(a).size === a.length;
-
-const validateBlankOrPositiveInteger = (n) => (!!n ? Number.isInteger(n) && n > 0 : true);
-
-const errorMessage = ({ path }) => {
-  const errorMessages = {
-    // Common fields
-    firstName: 'First name is required',
-    lastName: 'Last name is required',
-    phoneNumber: 'Phone number is required',
-    emailAddress: 'Email address is required',
-    postalCode: 'Postal code is required',
-
-    // Employer operator contact info
-    registeredBusinessName: 'Business name is required',
-    isRHO: 'Regional Health Office status is required',
-    operatorName: 'Operator name is required',
-    operatorContactFirstName: 'Operator contact first name is required',
-    operatorContactLastName: 'Operator contact last name is required',
-    operatorEmail: 'Operator email is required',
-    operatorPhone: 'Operator phone is required',
-    operatorAddress: 'Operator address is required',
-    operatorPostalCode: 'Operator postal code is required',
-
-    // Employer site contact info
-    siteName: 'Site name is required',
-    instituteName: 'Post Secondary Institute name is required',
-    address: 'Address is required',
-    streetAddress: 'Address is required',
-    healthAuthority: 'Health authority is required',
-    siteContactFirstName: 'First name is required',
-    siteContactLastName: 'Last name is required',
-
-    // Employer site type and size info
-    siteType: 'Site type is required',
-    numPublicLongTermCare: 'Number of publicly funded long-term care beds is required',
-    numPrivateLongTermCare: 'Number of privately funded long-term care beds is required',
-    numPublicAssistedLiving: 'Number of publicly funded assisted living beds is required',
-    numPrivateAssistedLiving: 'Number of privately funded assisted living beds is required',
-
-    // Workforce Baseline
-    workforceBaseline: 'All Workforce Baseline fields are required',
-
-    // Employer HCAP request
-    hcswFteNumber: 'A number is required',
-
-    // Employer certification
-    doesCertify: 'Must certify this',
-
-    // Orbeon ID from the XML file name
-    orbeonId: 'Invalid Orbeon ID format.',
-
-    // Participant info
-    eligibility:
-      "We're sorry, but current eligibility to work in Canada is a requirement to submit this form.",
-    preferredLocation: "Please select at least one location you'd like to work in.",
-    reasonForFindingOut: 'Please let us know how you found out about HCAP',
-    consent: "We're sorry, but we cannot process your request without permission.",
-
-    // Cohort
-    cohortName: 'Cohort Name is required',
-    startDate: 'Start Date is required',
-    endDate: 'End Date is required',
-    cohort: 'Cohort is required',
-    institute: 'Institute is required',
-  };
-  return errorMessages[path] || `Failed validation on ${path}`;
-};
+export const genericConfirm = yup.object().shape({
+  confirmed: yup.boolean().test('is-true', 'Please confirm', (v) => v === true),
+});
 
 export const LoginSchema = yup
   .object()
@@ -407,10 +308,6 @@ export const EditParticipantFormSchema = yup.object().shape({
     .matches(/^[A-Z]\d[A-Z]\s?\d[A-Z]\d$/, 'Invalid postal code'),
   emailAddress: yup.string().required(errorMessage).email('Invalid email address'),
   interested: yup.string().nullable(),
-});
-
-export const genericConfirm = yup.object().shape({
-  confirmed: yup.boolean().test('is-true', 'Please confirm', (v) => v === true),
 });
 
 export const ArchiveHiredParticipantSchema = yup.object().shape({
